tests/wpt/tests/portals/portal-onload-event.html

<!DOCTYPE html>
<title>Tests that the load is dispatched when a portal finishes loading.</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
  async_test(function(t) {
    assert_implements("HTMLPortalElement" in self);
    var w = window.open("resources/simple-portal.html");
    w.onload = function() {
      var portal = w.document.createElement("portal");
      portal.src = "resources/simple-portal.html";
      portal.onload = t.step_func_done();
      w.document.body.appendChild(portal);
    }
  }, "Tests that the load event is dispatched when a portal finishes loading.");
</script>
